Modularize ui.js into ui/ directory structure
Break monolithic lib/ui.js into modular components for easier tab additions:
- lib/ui/styles.js: CSS design system with dynamic author badges
- lib/ui/client-core.js: shared client JS (helpers, tabs, journal, status)
- lib/ui/shell.js: HTML page builder assembling all components
- lib/ui/tabs/github.js: GitHub tab client code (conditionally included)
- lib/ui.js: re-exports buildHTML for backward compatibility

Tab JS is now conditionally included — only enabled tabs get their JS
bundled into the HTML. TAB_LOADERS map replaces hardcoded switchTab
dispatch. Prepares for Phase 2 tab additions (roadmap, health, requests).

Refs #7
